pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "petey"
version = "0.3.1"
description = "Petey — The Easy PDF Extractor"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["pdf", "extraction", "llm", "openai", "anthropic", "structured-data"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing",
]
dependencies = [
    "PyMuPDF",
    "pymupdf4llm",
    "openai",
    "anthropic",
    "instructor",
    "pydantic",
    "pyyaml",
    "python-dotenv",
    "pdfplumber",
    "pytesseract",
    "Pillow",
    "litellm",
    "httpx",
    "python-dateutil",
]

[project.urls]
Homepage = "https://github.com/user/petey"
Repository = "https://github.com/user/petey"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "build",
    "twine",
]
docling = ["docling"]
liteparse = ["liteparse"]
unstructured = ["unstructured[pdf]"]
aws = ["boto3"]
google = ["google-cloud-documentai"]
azure = ["azure-ai-documentintelligence"]
all = [
    "petey[docling]",
    "petey[liteparse]",
    "petey[unstructured]",
    "petey[aws]",
    "petey[google]",
    "petey[azure]",
]

[tool.setuptools.packages.find]
include = ["petey*"]

[project.scripts]
petey = "petey.cli:main"
